Move command line utilities from CompileJob.swift to CommandLineArguments.swift
